Updated the HomeScreen.js
add common code file for reusable commponents
