+++ src/components/Screens/home/common/ConfirmationModal.js
@@ -1,0 +1,126 @@
+import React from 'react';
+import { View, Text, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
+import Icon from 'react-native-vector-icons/MaterialIcons';
+
+const { width } = Dimensions.get('window');
+
+const ConfirmationModal = ({
+  visible,
+  title,
+  message,
+  confirmText = 'CONFIRM',
+  cancelText = 'CANCEL',
+  onConfirm,
+  onCancel,
+  iconName = 'warning',
+  iconColor = '#FFD700'
+}) => {
+  if (!visible) return null;
+
+  return (
+    <View style={styles.modalContainer}>
+      <View style={styles.modalContent}>
+        <View style={styles.modalHeader}>
+          <Icon name={iconName} size={40} color={iconColor} />
+          <Text style={styles.modalTitle}>{title}</Text>
+        </View>
+        
+        <Text style={styles.modalText}>
+          {message}
+        </Text>
+        
+        <View style={styles.modalButtons}>
+          <TouchableOpacity 
+            style={[styles.modalButton, styles.cancelButton]}
+            onPress={onCancel}
+          >
+            <Text style={styles.modalButtonText}>{cancelText}</Text>
+          </TouchableOpacity>
+          
+          <TouchableOpacity 
+            style={[styles.modalButton, styles.confirmButton]}
+            onPress={onConfirm}
+          >
+            <Text style={styles.modalButtonText}>{confirmText}</Text>
+          </TouchableOpacity>
+        </View>
+      </View>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  modalContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.7)',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    zIndex: 1000,
+  },
+  modalContent: {
+    width: width * 0.8,
+    backgroundColor: '#1a1a2e',
+    borderRadius: 20,
+    padding: 25,
+    borderWidth: 3,
+    borderColor: '#FFD700',
+    shadowColor: '#FFD700',
+    shadowOffset: { width: 0, height: 0 },
+    shadowOpacity: 0.8,
+    shadowRadius: 10,
+    elevation: 10,
+  },
+  modalHeader: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginBottom: 20,
+  },
+  modalTitle: {
+    fontSize: 22,
+    fontWeight: 'bold',
+    color: '#FFD700',
+    marginLeft: 10,
+    textShadowColor: 'rgba(255, 215, 0, 0.5)',
+    textShadowOffset: { width: 1, height: 1 },
+    textShadowRadius: 5,
+  },
+  modalText: {
+    color: '#FFF',
+    fontSize: 16,
+    textAlign: 'center',
+    marginBottom: 25,
+  },
+  modalButtons: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  modalButton: {
+    flex: 1,
+    padding: 15,
+    borderRadius: 10,
+    alignItems: 'center',
+    marginHorizontal: 5,
+    borderWidth: 2,
+  },
+  cancelButton: {
+    backgroundColor: '#0f3460',
+    borderColor: '#2196F3',
+  },
+  confirmButton: {
+    backgroundColor: '#0f3460',
+    borderColor: '#F44336',
+  },
+  modalButtonText: {
+    color: '#FFF',
+    fontWeight: 'bold',
+    fontSize: 16,
+  },
+});
+
+export default ConfirmationModal;
